+++ frontend/src/app/state/application/selectors/exercise.selectors.ts
@@ -9,6 +9,7 @@ import type {
     ExerciseState,
     TechnicalChallengeId,
     UUID,
+    Vehicle,
     WithPosition,
 } from 'fuesim-digital-shared';
 import {
@@ -25,9 +26,10 @@ import type { TransferLine } from '../../../shared/types/transfer-line';
 /**
  * Check before via selectExerciseStateMode whether the exerciseState is defined
  */
-export const selectExerciseState = (state: AppState) =>
+export function selectExerciseState(state: AppState) {
     // TODO: we currently expect this to only be used of the exerciseStateMode is not undefined
-    state.application.exerciseState!;
+    return state.application.exerciseState!;
+}
 
 function selectPropertyFactory<Key extends keyof ExerciseState>(key: Key) {
     return createSelector(selectExerciseState, (exercise) => exercise[key]);
@@ -49,6 +51,9 @@ export const selectHospitalPatients = selectPropertyFactory('hospitalPatients');
 export const selectClients = selectPropertyFactory('clients');
 export const selectRadiograms = selectPropertyFactory('radiograms');
 export const selectRestrictedZones = selectPropertyFactory('restrictedZones');
+export const selectOperationalSections = selectPropertyFactory(
+    'operationalSections'
+);
 export const selectTechnicalChallenges = selectPropertyFactory(
     'technicalChallenges'
 );
@@ -139,6 +144,11 @@ export const selectTileMapProperties = createSelector(
     (configuration) => configuration.tileMapProperties
 );
 
+export const selectOperationsMapProperties = createSelector(
+    selectConfiguration,
+    (configuration) => configuration.operationsMapProperties
+);
+
 export const selectTransferLines = createSelector(
     selectExerciseState,
     selectTransferPoints,
@@ -191,6 +201,18 @@ export function createSelectReachableHospitals(transferPointId: UUID) {
     );
 }
 
+export function createSelectVehiclesInOperationalSection(
+    operationalSectionId: UUID
+) {
+    return createSelector(selectVehicles, (vehicles) =>
+        Object.values(vehicles).filter(
+            (vehicle) =>
+                vehicle.operationalAssignment?.type === 'operationalSection' &&
+                vehicle.operationalAssignment.sectionId === operationalSectionId
+        )
+    );
+}
+
 export const selectVehiclesInTransfer = createSelector(
     selectVehicles,
     (vehicles) =>
@@ -202,6 +224,77 @@ export const selectPersonnelInTransfer = createSelector(
     (personnel) =>
         Object.values(personnel).filter((_personnel) =>
             isInTransfer(_personnel)
+        )
+);
+
+export const selectLocalOperationsCommand = createSelector(
+    selectVehicles,
+    (vehicles) =>
+        Object.values(vehicles).find(
+            (v) => v.operationalAssignment?.type === 'localOperationsCommand'
+        )
+);
+
+export function createSelectOperationalSectionLeader(sectionId: string) {
+    return createSelector(
+        createSelectVehiclesInOperationalSection(sectionId),
+        (vehicles) =>
+            Object.values(vehicles)
+                .filter(
+                    (vehicle) =>
+                        vehicle.operationalAssignment?.type ===
+                            'operationalSection' &&
+                        vehicle.operationalAssignment.role ===
+                            'operationalSectionLeader'
+                )
+                .at(0)
+    );
+}
+
+export function createSelectOperationalSectionMembers(sectionId: string) {
+    return createSelector(
+        createSelectVehiclesInOperationalSection(sectionId),
+        (vehicles) =>
+            Object.values(vehicles).filter(
+                // This type annotation is necessary to not have to repeat the same conditions later
+                (
+                    vehicle
+                ): vehicle is Vehicle & {
+                    operationalAssignment: NonNullable<
+                        Vehicle['operationalAssignment']
+                    > & {
+                        role: 'operationalSectionMember';
+                        position: number;
+                    };
+                } =>
+                    vehicle.operationalAssignment?.type ===
+                        'operationalSection' &&
+                    vehicle.operationalAssignment.role ===
+                        'operationalSectionMember'
+            )
+    );
+}
+
+export function createSelectSortedOperationalSectionMembers(sectionId: string) {
+    return createSelector(
+        createSelectOperationalSectionMembers(sectionId),
+        (vehicles) =>
+            vehicles.sort(
+                (a, b) =>
+                    a.operationalAssignment.position -
+                    b.operationalAssignment.position
+            )
+    );
+}
+
+export const selectVehiclesInTransferFromAlarmgroup = createSelector(
+    selectVehiclesInTransfer,
+    (vehicles) =>
+        Object.values(vehicles).filter(
+            (vehicle) =>
+                vehicle.position.type === 'transfer' &&
+                vehicle.position.transfer.startPoint.type ===
+                    'alarmGroupStartPoint'
         )
 );