+++ shared/src/state-migrations/migration-functions.ts
@@ -45,6 +45,7 @@ import { generalizeMaterialsPersonnel44 } from './44-generalize-materials-person
 import { addRestrictedZones45 } from './45-add-restricted-zones.js';
 import { limitedAlarmgroups46 } from './46-limited-alarm-groups.js';
 import { participantIdToKey47 } from './47-participant-id-to-key.js';
+import { addOperationsTabletView49 } from './49-add-operations-tablet-view.js';
 import { addAutojoinViewport48 } from './48-autojoin-viewport.js';
 
 /**
@@ -120,4 +121,5 @@ export const migrations: {
     46: limitedAlarmgroups46,
     47: participantIdToKey47,
     48: addAutojoinViewport48,
+    49: addOperationsTabletView49,
 };